pets integrated into server

+++ server/mongodb/operations/address_operations.js
@@ -1,24 +1,29 @@
 ﻿const {Address} = require("../schemas/address_schema");
+const {isValidObjectId} = require("../../utils/database_utils");
 
 module.exports.findAddress = async (id) => {
+    if (!isValidObjectId(id)) return false
     return Address.findOne({
         userId: id
     });
 };
 
 module.exports.doesAddressExist = async (id) => {
+    if (!isValidObjectId(id)) return false
     return Address.exists({
         userId: id
     });
 };
 
 module.exports.deleteAddress = async (id) => {
+    if (!isValidObjectId(id)) return false
     await Address.findOneAndRemove({
         userId: id
     });
 };
 
 module.exports.createAddress = async (id, address) => {
+    if (!isValidObjectId(id)) return false
     address.userId = id;
     const newAddress = await new Address(address);
     await newAddress.save();
@@ -26,6 +31,7 @@ module.exports.createAddress = async (id, address) => {
 };
 
 module.exports.updateAddress = async (id, address) => {
+    if (!isValidObjectId(id)) return false
     await Address.findOneAndUpdate({
             id: id
         }, address
